pfp: separate recorders options
Place Recorder and NodeRecorder options in one file for easier
navigation.
